+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # @project-version@ (@build-time@)
 
+* Fixed a Lua error that occured sometimes when friendly pets/totems/guardians turn neutral, e.g. when losing or giving up control over them [Issue #511].
 * Updated TOC version for for TBC Classic Patch 2.5.2.